# openapi.yaml
openapi: 3.1.0
info:
  title: PM Assistant API
  version: 1.0.0
  description: >
    REST API สำหรับผู้ช่วย PM บน LINE — ใช้โดยหน้าเว็บตั้งค่า (LIFF) และรับ event จาก
    LINE/แพลตฟอร์มภายนอก. ใช้คู่กับ schema.sql และ DESIGN_SPEC.md.
    ทุก endpoint ฝั่ง LIFF ต้องมี bearer token (จาก LINE Login); endpoint /webhooks
    ตรวจ signature แทน.
servers:
  - url: https://api.pmassistant.app/v1
    description: production
  - url: http://localhost:8000/v1
    description: local dev
...
security:
  - bearerAuth: []
paths:
...
  /projects/{projectId}/members:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Members]
      summary: รายชื่อสมาชิก
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Member' }
    post:
      tags: [Members]
      summary: เพิ่มสมาชิก
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberInput' }
      responses:
        '201':
          description: สำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
  /members/{memberId}:
    parameters:
      - name: memberId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Members]
      summary: แก้ไขสมาชิก
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberInput' }
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
    delete:
      tags: [Members]
      summary: ลบสมาชิก
      responses:
        '204': { description: สำเร็จ }

  # ------------------------------------------------------------ Connections
...
  /connections/authorize:
    post:
      tags: [Connections]
      summary: เริ่มเชื่อมต่อแพลตฟอร์ม (คืน URL หน้าอนุญาตสิทธิ์)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: { $ref: '#/components/schemas/ConnectionType' }
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  authorizationUrl: { type: string, format: uri }
  /connections/oauth/callback:
    get:
      tags: [Connections]
      summary: callback ของ OAuth (แลก code เป็น token แล้วบันทึก connection)
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          required: true
          schema: { type: string }
      responses:
        '302': { description: redirect กลับหน้า LIFF }
  /connections/{connectionId}:
    parameters:
      - name: connectionId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [Connections]
      summary: ยกเลิกการเชื่อมต่อ
      responses:
        '204': { description: สำเร็จ }

  # --------------------------------------------------------------- Bindings
  /projects/{projectId}/bindings:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Bindings]
      summary: ปลายทางของแต่ละ capability
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Binding' }
  /projects/{projectId}/bindings/{capability}:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - name: capability
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Capability' }
    put:
      tags: [Bindings]
      summary: ตั้งว่า capability นี้ใช้ connection ไหน
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [connectionId]
              properties:
                connectionId: { type: string, format: uuid }
                config: { type: object, additionalProperties: true }
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Binding' }

  # -------------------------------------------------------------- Templates
...
  /projects/{projectId}/documents:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - name: wpType
        in: query
        schema: { $ref: '#/components/schemas/WorkProductType' }
      - $ref: '#/components/parameters/limit'
      - $ref: '#/components/parameters/cursor'
    get:
      tags: [Documents]
      summary: คลังเอกสาร (หน้า W8)
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Document' }
                  nextCursor: { type: string, nullable: true }
  /documents/{documentId}/download:
    parameters:
      - name: documentId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Documents]
      summary: ลิงก์ดาวน์โหลดเอกสาร (signed URL)
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  downloadUrl: { type: string, format: uri }

  # ------------------------------------------------------------------ Tasks
...
  /projects/{projectId}/meetings:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Meetings]
      summary: รายการบันทึกการประชุม
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Meeting' }

  # --------------------------------------------------------------- Calendar
  /projects/{projectId}/events:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Calendar]
      summary: รายการนัดหมาย
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CalendarEvent' }
    post:
      tags: [Calendar]
      summary: สร้างนัดหมาย (recurrence + Meet link + invite)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CalendarEventInput' }
      responses:
        '201':
          description: สำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CalendarEvent' }

  # ----------------------------------------------------------- Traceability
...
  /confirmations/{confirmationId}/confirm:
    parameters:
      - name: confirmationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Confirmations]
      summary: ยืนยันร่างเพื่อ execute (สะท้อน postback จาก LINE)
      responses:
        '200':
          description: ผลการ execute
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutionResult' }
        '409':
          description: ร่างหมดอายุหรือถูกจัดการไปแล้ว
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /confirmations/{confirmationId}/cancel:
    parameters:
      - name: confirmationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Confirmations]
      summary: ยกเลิกร่าง
      responses:
        '204': { description: สำเร็จ }

  # ----------------------------------------------------------- Notifications
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: token ที่ออกหลังตรวจ LINE Login ID token จากฝั่ง LIFF
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    cursor:
      name: cursor
      in: query
      schema: { type: string }
...
  schemas:
...
    Capability:
      type: string
      enum: [tasks, calendar, docs, email, notify]
    ConnectionType:
      type: string
      enum: [jira, clickup, google, gmail, line, other]
    WorkProductType:
      type: string
      enum: [meeting_record, memo, project_plan, requirements, traceability, test_case, change_request]
...
    DocStatus:
      type: string
      enum: [draft, issued]

    # ---- error ----
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: not_found }
        message: { type: string }
        details: { type: object, additionalProperties: true }

    # ---- projects ----
...
    Member:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email, nullable: true }
        role: { type: string, nullable: true }
    MemberInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string }

    # ---- connections / bindings ----
...
    Binding:
      type: object
      properties:
        id: { type: string, format: uuid }
        capability: { $ref: '#/components/schemas/Capability' }
        connectionId: { type: string, format: uuid }
        config: { type: object, additionalProperties: true }

    # ---- templates / numbering ----
...
    Document:
      type: object
      properties:
        id: { type: string, format: uuid }
        wpType: { $ref: '#/components/schemas/WorkProductType' }
        docNumber: { type: string, nullable: true, example: HR-MIN-0007 }
        title: { type: string }
        status: { $ref: '#/components/schemas/DocStatus' }
        createdAt: { type: string, format: date-time }

    # ---- tasks / milestones ----
...
    Meeting:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        meetingDate: { type: string, format: date, nullable: true }
        decisions:
          type: array
          items: { type: string }
        actionItems:
          type: array
          items:
            type: object
            properties:
              description: { type: string }
              ownerId: { type: string, format: uuid, nullable: true }
              dueDate: { type: string, format: date, nullable: true }
              isInferred: { type: boolean, description: ระบบเดา }

    # ---- calendar ----
    CalendarEvent:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time, nullable: true }
        recurrenceRule: { type: string, nullable: true, example: 'FREQ=WEEKLY;BYDAY=MO' }
        meetLink: { type: string, format: uri, nullable: true }
    CalendarEventInput:
      type: object
      required: [title, startsAt]
      properties:
        title: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        recurrenceRule: { type: string }
        attendeeMemberIds:
          type: array
          items: { type: string, format: uuid }
        createMeetLink: { type: boolean, default: true }

    # ---- traceability ----
...
    ExecutionResult:
      type: object
      properties:
        status: { type: string, enum: [success, partial, failed] }
        actions:
          type: array
          items:
            type: object
            properties:
              tool: { type: string, example: clickup }
              ok: { type: boolean }
              ref: { type: string, nullable: true, example: HR-128 }
              error: { type: string, nullable: true }

    # ---- notifications ----
...